Switch to new metadata; implement Open in Bloom button
(Note: this requires the version of bloom desktop from the BloomLibraryIntegration branch, checkin 364eec60b1c964f7de334783d22597d38f146aba or a descendant.
For the Open in Bloom button to work, registry entries must be created by running the installer or, on  a dev machine, editing and using the registry changes
in "bloom link.reg" in the root directory. Books to be downloaded in this way must also have been uploaded by that version. Also, for now, books to be
downloaded by a debug build must have been uploaded by one. Currently the books are in different buckets on amazon but share one project on parse.com).