Make the detail view dialog disable the rest of the page while it is active
This used to work only after a Refresh. I think the functionality was disrupted when we changed to use a distinct URL for each book displayed.
The problem turned out to be a bug in angular-bootstrap's Dialog module, so the fix involves using a patched version of the vendor javascript.