Sticky footer with terms of use and privacy
index.html diff is confusing. What really happened is, whole former body is wrapped in new
div.page-wrap, and div.site-footer is added below that. Terms of use and privacy links moved there
(and no longer need to be duplicated).

Since terms and privacy can now be invoked when the page is scrolled, added some code to
scroll to top when they are activated.