Improve logo
The previous logo had the problem that while Serif is available on
Linux it renders differently from Windows, resulting in "let's grow a
library" slogan overlapping "Bloom". This change makes use of the CSS
style for specifying the font (see
http://www.w3.org/TR/SVG/styling.html#StyleAttribute) and specifies
multiple fonts which are tried in the given order. Unfortunately
Inkscape in it's current version isn't terribly good in supporting
fallback fonts, but at least it seems to accept them.

This change is based on the original SVG file from commit 9493069.