Handle case where a book points to a deleted language
We had a problem where many books list languages that they don't actually have content for. So as to not falsely advertise having books in those languages, we delete the rows by hand, for now. When that happens, we need this code so that those books can be displayed without error.