Datagrid auto-adjusts to full browser window height
It was necessary to use an event listener on the window resize instead of using CSS 100% because the datagrid has no control over its parents' dimensions.

I used a variable outside of $scope in the controller file because I could not figure out a way to allow the onExit function of the stateProvider to access $scope. It is necessary to remove the event listener since it is attached to the window.