Added download history with column in datagrid
When a user downloads a book, a request is sent to icanhazip.com to get the user's ip address. The ip, bookId, and username (and date by default) are then saved in an entry on the downloadHistory table on parse.com. After the entry is saved, the book's downloadCount is incremented via cloud-code aftersave on downloadHistory.