Added Related Books Functionality
Datagrid has added column on the right with "Relate" links. Clicking on one of these pulls up the existing relationship the book may or may not already have with other books.
Books may be added to this relation via autocomplete or may be removed. If a book is added that is already a part of a relationship, a dialog box appears asking if the user is sure he/she to relate all of these books.
Once a relationship is established, related books show up as a comma-separated list of links in the detail view.

- Added relatedBook table  to parse.com
- Added manipulation of the table in bookService
- Added related books modal
- Added confirmation modal (for already related books)